+++ packages/@ionic/cli-framework-output/src/logger.ts
@@ -1,12 +1,9 @@
-import * as chalk from 'chalk';
-import * as lodash from 'lodash';
 import { Writable } from 'stream';
 import * as util from 'util';
 
-import { WordWrapOptions, stringWidth, wordWrap } from '../utils/format';
-import { enforceLF } from '../utils/string';
-
-import { ColorFunction, Colors, DEFAULT_COLORS } from './colors';
+import { ColorFunction, Colors, NO_COLORS } from './colors';
+import { WordWrapOptions, stringWidth, wordWrap } from './format';
+import { dropWhile, enforceLF } from './utils';
 
 export interface LogRecord {
   msg: string;
@@ -221,29 +218,47 @@ export interface CreateTaggedFormatterOptions {
   titleize?: boolean;
   wrap?: boolean | WordWrapOptions;
   colors?: Colors;
+  tags?: ReadonlyMap<LoggerLevelWeight, string>;
 }
 
-export function createTaggedFormatter({ colors = DEFAULT_COLORS, prefix = '', titleize, wrap }: CreateTaggedFormatterOptions = {}): LoggerFormatter {
+export function createTaggedFormatter({ colors = NO_COLORS, prefix = '', tags, titleize, wrap }: CreateTaggedFormatterOptions = {}): LoggerFormatter {
+  const { strong, weak } = colors;
+
+  const getLevelTag = (level?: LoggerLevelWeight): string => {
+    if (!level) {
+      return '';
+    }
+
+    if (tags) {
+      const tag = tags.get(level);
+
+      return tag ? tag : '';
+    }
+
+    const levelName = getLoggerLevelName(level);
+
+    if (!levelName) {
+      return '';
+    }
+
+    const levelColor = getLoggerLevelColor(colors, level);
+
+    return `${weak('[')}\x1b[40m${strong(levelColor ? levelColor(levelName) : levelName)}\x1b[49m${weak(']')}`;
+  };
+
   return ({ msg, level, format }) => {
     if (format === false) {
       return msg;
     }
 
-    const { strong, weak } = colors;
-
     const [ firstLine, ...lines ] = msg.split('\n');
 
-    const levelName = getLoggerLevelName(level);
     const levelColor = getLoggerLevelColor(colors, level);
 
-    const tag = (
-      (typeof prefix === 'function' ? prefix() : prefix) +
-      (levelName ? `${weak('[')}${chalk.bgBlack(strong(levelColor ? levelColor(levelName) : levelName))}${weak(']')}` : '')
-    );
-
+    const tag = (typeof prefix === 'function' ? prefix() : prefix) + getLevelTag(level);
     const title = titleize && lines.length > 0 ? `${strong(levelColor ? levelColor(firstLine) : firstLine)}\n` : firstLine;
     const indentation = tag ? stringWidth(tag) + 1 : 0;
-    const pulledLines = lodash.dropWhile(lines, l => l === '');
+    const pulledLines = dropWhile(lines, l => l === '');
 
     return (
       (tag ? `${tag} ` : '') +