+++ packages/@ionic/cli/src/lib/shell.ts
@@ -1,4 +1,4 @@
-import { LOGGER_LEVELS } from '@ionic/cli-framework';
+import { LOGGER_LEVELS } from '@ionic/cli-framework-output';
 import { createProcessEnv, killProcessTree, onBeforeExit } from '@ionic/utils-process';
 import { ERROR_COMMAND_NOT_FOUND, Subprocess, SubprocessError, SubprocessOptions, WhichOptions, which } from '@ionic/utils-subprocess';
 import { TERMINAL_INFO } from '@ionic/utils-terminal';